% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepp-package.R
\name{movepp_version}
\alias{movepp_version}
\title{Return movepp version}
\usage{
movepp_version()
}
\value{
A \code{package_version} object.
}
\description{
Convenience function returning the installed package version.
}
\examples{
movepp_version()
}
