% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi.R
\name{print.movepp_dominant_axis}
\alias{print.movepp_dominant_axis}
\title{Print method for movepp_dominant_axis objects}
\usage{
\method{print}{movepp_dominant_axis}(x, ...)
}
\arguments{
\item{x}{A \code{movepp_dominant_axis} object from \code{\link[=detect_dominant_axis]{detect_dominant_axis()}}.}

\item{...}{Ignored.}
}
\description{
Print method for movepp_dominant_axis objects
}
